fix: unmarshal number leads to int overflow
Resolves https://github.com/maas/terraform-provider-maas/issues/116

+++ entity/block_device_partition.go
@@ -3,7 +3,7 @@ package entity
 // BlockDevicePartition represents the MaaS block device partition endpoint.
 type BlockDevicePartition struct {
 	UUID        string              `json:"uuid"`
-	Size        int                 `json:"size"`
+	Size        int64               `json:"size"`
 	Bootable    bool                `json:"bootable"`
 	Tags        []string            `json:"tags,omitempty"`
 	SystemID    string              `json:"system_id"`
@@ -25,7 +25,7 @@ type PartitionFileSystem struct {
 }
 
 type BlockDevicePartitionParams struct {
-	Size     int    `url:"size,omitempty"`
+	Size     int64  `url:"size,omitempty"`
 	UUID     string `url:"uuid,omitempty"`
 	Bootable bool   `url:"bootable"`
 }
